+++ src/actions/blogs.ts
@@ -2,26 +2,27 @@
 
 import { cacheLife, cacheTag } from 'next/cache';
 import { Data as portfolioData } from '@/data/portfolio-data';
-import { SITE_URL } from '@/lib/constants';
+import { getMediumPost, scrapeWebsite } from '@/lib/blogs-utils';
 import type { Blog } from '@/types/portfolio-types';
 
 const getBlogs = async () => {
   'use cache';
-  cacheTag('blogs');
   cacheLife('days');
+  cacheTag('blogs');
+
   try {
     const blogsPromise = portfolioData.blogs.sources.map(async (source) => {
+      let blogs: Blog[] = [];
       if (source.source === 'Medium') {
-        const response = await fetch(`${SITE_URL}/api/blogs?username=${source.username}`);
-        return response.json() as Promise<Blog[]>;
+        blogs = await getMediumPost(source.username!);
       } else {
         const customBlogsPromise = source.urls?.map(async (url) => {
-          const response = await fetch(`${SITE_URL}/api/blogs?url=${url}`);
-          return response.json() as Promise<Blog[]>;
+          const otherBlogs = await scrapeWebsite(url);
+          return otherBlogs;
         });
-        const customBlogs = await Promise.all(customBlogsPromise || []);
-        return customBlogs.flat();
+        blogs = (await Promise.all(customBlogsPromise || [])).flat();
       }
+      return blogs;
     });
     const blogs = await Promise.all(blogsPromise);
     return blogs.flat();
