+++ src/actions/videos.ts
@@ -2,18 +2,19 @@
 
 import { cacheLife, cacheTag } from 'next/cache';
 import { Data as portfolioData } from '@/data/portfolio-data';
-import { SITE_URL } from '@/lib/constants';
+import { fetchChannelStats, fetchYouTubeVideos } from '@/lib/videos-utils';
 
 const getVideos = async () => {
   'use cache';
-  cacheTag('videos');
   cacheLife('days');
+  cacheTag('videos');
 
   const channelId = portfolioData.videos.youtubeChannelId;
   if (!channelId) return;
-  const res = await fetch(`${SITE_URL}/api/videos?channelId=${channelId}`);
-  if (!res.ok) return;
-  return await res.json();
+
+  const videos = await fetchYouTubeVideos({ channelId });
+  const channelStats = await fetchChannelStats(channelId);
+  return { videos, channelStats };
 };
 
 export { getVideos };
