+++ src/app/(resources)/videos/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next';
 import { Suspense } from 'react';
-import VideoContent from '@/components/pages/videos-content';
+import VideosServerContent from '@/components/pages/videos-server-content';
 import { VideosSkeleton } from '@/components/skeletons/videos';
 
 export const metadata: Metadata = {
@@ -11,7 +11,7 @@ export const metadata: Metadata = {
 export default function VideosPage() {
   return (
     <Suspense fallback={<VideosSkeleton />}>
-      <VideoContent />
+      <VideosServerContent />
     </Suspense>
   );
 }
